Rewrite inline help to support keywords
This is a redesign of the feature with a clear architecture. It not only handles keywords, it should also work with generic UTF-8 identifiers. A lot more test cases are added, where a large amount is auto-generated on thy fly.

If cases happen which are not explicitly covered, we at least try to do something useful. Nevertheless, this certainly needs to be improved with more test cases coming up.