🩹 Deactivate test with known JuliaSyntax.jl bug
This throws on nightly so that it is harder to see real problems in TerminalPager.jl. The problem should be fixed in JuliaSyntax.jl, so is not fully actionable in TerminalPager.jl.

This should be reverted when [https://github.com/JuliaLang/JuliaSyntax.jl/issues/599](https://github.com/JuliaLang/JuliaSyntax.jl/issues/599) is fixed.