🔧 Implement more robust shortcut registration
For all I know the previous code is currently working as intended. However, strictly speaking, we do not know which indices REPL uses. We are only interested in the first and the last element, so now the code matches the documentation. While we are at it, we can avoid repeating ourself and highlight the fact that we are operating on only one iterable.